Refactor Format base class
