+++ brainglobe/citation/bibtex_fmt.py
@@ -1,11 +1,12 @@
 import inspect
 import sys
-import warnings
 from string import ascii_letters, digits
-from typing import Any, ClassVar, Dict, List
+from typing import Any, ClassVar, Dict
+
+from brainglobe.citation.format import Format
 
 
-class BibTexEntry:
+class BibTexEntry(Format):
     """
     An abstract base class for generating BibTex entries from
     CITATION.cff yaml-content.
@@ -29,33 +30,10 @@ class BibTexEntry:
     ----------
     cite_key : str
         The citation key that appears in the BibTex reference
-    required : ClassVar[List[str]]
-        Required fields for the entry to be created
-    optional : ClassVar[List[str]]
-        Optional fields that the entry may possess
-    indent_character : str, default = ' ' * 4
-        Character sequence to use for indentation.
     """
 
-    # The citation key that appears in the BibTex reference
+    author_separator: ClassVar[str] = " and "
     cite_key: str
-    # Required fields for the entry to be created
-    required: ClassVar[List[str]]
-    # Optional fields that the entry may possess
-    optional: ClassVar[List[str]]
-
-    # Character sequence to use for indentation
-    indent_character: str = " " * 4
-
-    # mypy type-hints
-    authors: str | Dict[str, str] | List[Dict[str, str]]
-
-    @classmethod
-    def entry_type(cls) -> str:
-        """
-        The Bibtex entry type; article, book, software, etc.
-        """
-        return cls.__name__.lower()
 
     @classmethod
     def validate_citation_key(cls, key: str) -> bool:
@@ -114,75 +92,8 @@ class BibTexEntry:
             # so we don't try to assign it to a field later.
             information.pop("type")
 
-        # Add all the information we need
-        for key, value in information.items():
-            if key in self.required or key in self.optional:
-                setattr(self, key, value)
-            elif warn_on_not_used:
-                warnings.warn(
-                    f"The key {key} is not used for entries of type "
-                    f"{self.entry_type()}",
-                    UserWarning,
-                )
+        super().__init__(information, warn_on_not_used=warn_on_not_used)
 
-        # Check that all required information is populated
-        for required_field in self.required:
-            if not hasattr(self, required_field):
-                raise KeyError(
-                    f"Did not receive value for required key: {required_field}"
-                )
-        # Optional fields should be set to None so that checks against
-        # them produce nothing and evaluated to False
-        for optional_field in self.optional:
-            if not hasattr(self, optional_field):
-                setattr(self, optional_field, None)
-
-        # If we have an authors field, we will need to parse the
-        # dictionary input into the string that BibTex is expecting
-        if hasattr(self, "authors"):
-            self._prepare_authors_field()
-
-        return
-
-    def _prepare_authors_field(self) -> None:
-        """
-        The authors field may come in as a dict, or a list of dicts.
-        This is rather inconvenient as we need it to be a string, so
-        convert it with this function.
-
-        Individual author fields in CITATION.cff are:
-        - family-names
-        - given-names
-        - orcid
-        - affiliation
-
-        or which we only need the names.
-        """
-        # A single author will be read in as a dictionary
-        if isinstance(self.authors, dict):
-            surname = self.authors["family-names"]
-            forename = self.authors["given-names"]
-            self.authors = f"{forename} {surname}"
-        # Multiple authors will be read in as a list of dictionaries
-        elif isinstance(self.authors, list):
-            all_authors = []
-            for author_info in self.authors:
-                if not isinstance(author_info, dict):
-                    raise ValueError(
-                        f"Expected authors to be dictionary but it was"
-                        f" {type(author_info)}: {author_info}"
-                    )
-                else:
-                    surname = author_info["family-names"]
-                    forename = author_info["given-names"]
-                    all_authors.append(f"{forename} {surname}")
-            self.authors = " and ".join(all_authors)
-        # Unrecognised read format, abort
-        else:
-            raise TypeError(
-                f"Expected authors to be either dict or list of dicts,"
-                f" not {type(self.authors).__name__}"
-            )
         return
 
     def generate_ref_string(self) -> str:
@@ -267,7 +178,7 @@ class Software(BibTexEntry):
     ]
 
 
-def supported_entry_types() -> Dict[str, BibTexEntry]:
+def supported_bibtex_entry_types() -> Dict[str, BibTexEntry]:
     """
     Create a dict of all the classes in this module that can be used
     to write a bibtex reference of a particular entry type.
@@ -287,7 +198,7 @@ def supported_entry_types() -> Dict[str, BibTexEntry]:
         for name, obj in inspect.getmembers(
             sys.modules[__name__], inspect.isclass
         )
-        if name != "BibTexEntry"
+        if name != "BibTexEntry" and name != "Format"
     }
 
     return dict_of_formats
